% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{attrition}
\alias{attrition}
\title{Cohort attrition}
\usage{
attrition(x)
}
\arguments{
\item{x}{Cohort set for which to get attrition}
}
\value{
tibble with counts and reasons for attrition.
}
\description{
Cohort attrition
}
