% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{availableRollingIncidenceGrouping}
\alias{availableRollingIncidenceGrouping}
\title{Variables that can be used for faceting and colouring rolling incidence plots}
\usage{
availableRollingIncidenceGrouping(result, varying = FALSE)
}
\arguments{
\item{result}{Rolling incidence results}

\item{varying}{If FALSE, only variables with non-unique values will be
returned, otherwise all available variables will be returned}
}
\description{
Variables that can be used for faceting and colouring rolling incidence plots
}
