% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarkIncidencePrevalence.R
\name{benchmarkIncidencePrevalence}
\alias{benchmarkIncidencePrevalence}
\title{Run benchmark of incidence and prevalence analyses}
\usage{
benchmarkIncidencePrevalence(
  cdm,
  sample = NULL,
  outputFolder = NULL,
  verbose = FALSE
)
}
\arguments{
\item{cdm}{A CDM reference object}

\item{sample}{An integer for which to take a random sample when generating
the denominator cohort}

\item{outputFolder}{Folder to save results as CSV}

\item{verbose}{Either TRUE or FALSE. If TRUE, progress will be reported.}
}
\value{
a tibble with time taken for different analyses
}
\description{
Run benchmark of incidence and prevalence analyses
}
\examples{
\dontrun{
con <- DBI::dbConnect(duckdb::duckdb(), CDMConnector::eunomia_dir())
cdm <- CDMConnector::cdm_from_con(
  con = con,
  cdm_schema = "main"
)

timings <- IncidencePrevalence::benchmarkIncidencePrevalence(cdm)
}
}
