% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cohortCount}
\alias{cohortCount}
\title{Counts of cohorts in a cohort set}
\usage{
cohortCount(cohortTable, cohortDefinitionId = NULL)
}
\arguments{
\item{cohortTable}{Set of cohorts}

\item{cohortDefinitionId}{cohortDefinitionId can be used to
specify a particular cohort of interest}
}
\value{
tibble with settings used when generating the cohort set
}
\description{
Counts of cohorts in a cohort set
}
\examples{
cdm <- mockIncidencePrevalenceRef(sampleSize = 200)
cdm$denominator <- generateDenominatorCohortSet(cdm,
  startDate = NULL,
  endDate = NULL,
  ageGroup = list(c(18, 40)),
  sex = c("Female", "Male"),
  daysPriorHistory = 120
)
cohortCount(cohortTable = cdm$denominator)
cohortCount(cohortTable = cdm$denominator, cohortDefinitionId = 1)
}
