% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_pop_incidence.R
\name{collect_pop_incidence}
\alias{collect_pop_incidence}
\title{Collect population incidence estimates}
\usage{
collect_pop_incidence(
  cdm,
  table_name_denominator,
  table_name_outcomes,
  cohort_ids_denominator_pops = NULL,
  cohort_ids_outcomes = NULL,
  time_interval = "Months",
  outcome_washout_windows = 0,
  repetitive_events = FALSE,
  confidence_interval = "poisson",
  minimum_cell_count = 5,
  verbose = FALSE
)
}
\arguments{
\item{cdm}{CDMConnector CDM reference object}

\item{table_name_denominator}{table_name_denominator}

\item{table_name_outcomes}{Name of the table with the outcome cohorts}

\item{cohort_ids_denominator_pops}{Cohort ids of denominator populations}

\item{cohort_ids_outcomes}{Outcome cohort ids}

\item{time_interval}{Time intervals for incidence estimates}

\item{outcome_washout_windows}{Clean windows}

\item{repetitive_events}{Repeated events}

\item{confidence_interval}{Method for confidence intervals}

\item{minimum_cell_count}{Minimum number of events to report- results
lower than this will be obscured. If NULL all results will be reported.}

\item{verbose}{Whether to report progress}
}
\description{
Collect population incidence estimates
}
