% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_pop_point_prevalence.R
\name{collect_pop_point_prevalence}
\alias{collect_pop_point_prevalence}
\title{Collect population point prevalence estimates}
\usage{
collect_pop_point_prevalence(
  cdm_ref,
  table_name_outcomes,
  cohort_ids_outcomes,
  study_denominator_pop,
  cohort_ids_denominator_pops,
  time_intervals = "months",
  points = "start",
  confidence_interval = "none",
  minimum_cell_count = 5,
  verbose = FALSE
)
}
\arguments{
\item{cdm_ref}{CDMConnector CDM reference object}

\item{table_name_outcomes}{Name of the table with the outcome cohorts}

\item{cohort_ids_outcomes}{Outcome cohort ids}

\item{study_denominator_pop}{Tibble with denominator populations}

\item{cohort_ids_denominator_pops}{Cohort ids of denominator populations}

\item{time_intervals}{Time intervals for prevalence estimates}

\item{points}{where to compute the point prevalence}

\item{confidence_interval}{Method for confidence intervals}

\item{minimum_cell_count}{Minimum number of events to report- results
lower than this will be obscured. If NULL all results will be reported.}

\item{verbose}{Whether to report progress}
}
\value{

}
\description{
Collect population point prevalence estimates
}
