% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_pop_prevalence.R
\name{collect_pop_prevalence}
\alias{collect_pop_prevalence}
\title{Collect population prevalence estimates}
\usage{
collect_pop_prevalence(
  cdm,
  table_name_denominator,
  cohort_ids_denominator_pops,
  table_name_outcomes,
  cohort_ids_outcomes,
  type = "point",
  time_intervals = "months",
  full_period_requireds = TRUE,
  points = "start",
  minimum_representative_proportions = 0.5,
  confidence_interval = "binomial",
  minimum_cell_count = 5,
  verbose = FALSE
)
}
\arguments{
\item{cdm}{CDMConnector CDM reference object}

\item{table_name_denominator}{table_name_denominator}

\item{cohort_ids_denominator_pops}{Cohort ids of denominator populations}

\item{table_name_outcomes}{Name of the table with the outcome cohorts}

\item{cohort_ids_outcomes}{Outcome cohort ids}

\item{type}{type of prevalence, point or period}

\item{time_intervals}{Time intervals for prevalence estimates}

\item{full_period_requireds}{If full period is required}

\item{points}{where to compute the point prevalence}

\item{minimum_representative_proportions}{Minimum proportions that
individuals must have to contribute}

\item{confidence_interval}{Method for confidence intervals}

\item{minimum_cell_count}{Minimum number of events to report- results
lower than this will be obscured. If NULL all results will be reported.}

\item{verbose}{Whether to report progress}
}
\description{
Collect population prevalence estimates
}
