% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectDenominator.R
\name{collectDenominator}
\alias{collectDenominator}
\title{Identify a set of denominator populations}
\usage{
collectDenominator(
  cdm,
  startDate = NULL,
  endDate = NULL,
  ageStrata = list(c(0, 150)),
  sexStrata = "Both",
  daysPriorHistory = 0,
  strataTable = NULL,
  strataId = NULL,
  sample = NULL,
  verbose = FALSE
)
}
\arguments{
\item{cdm}{CDMConnector CDM reference object}

\item{startDate}{Date indicating the start of the study
period. If NULL,
the earliest observation_start_date in the observation_period table
will be used.}

\item{endDate}{Date indicating the end of the study
period. If NULL,
the latest observation_end_date in the observation_period table
will be used.}

\item{ageStrata}{List of age groups}

\item{sexStrata}{Sex of the cohorts}

\item{daysPriorHistory}{Days of prior history required to enter
the study cohort.}

\item{strataTable}{strataTable}

\item{strataId}{strata cohort definition id}

\item{sample}{sample n}

\item{verbose}{Either TRUE or FALSE.
If TRUE, progress will be reported.}
}
\description{
\code{collectDenominator()} creates a set of denominator cohorts
}
\examples{
\dontrun{
db <- DBI::dbConnect(" Your database connection here ")
cdm <- CDMConnector::cdm_from_con(
  con = db,
  cdm_schema = "cdm schema name"
)
dpop <- collectDenominator(
  cdm = cdm,
  startDate = as.Date("2008-01-01"),
  endDate = as.Date("2018-01-01")
)
}
}
