% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectPopIncidence.R
\name{collectPopIncidence}
\alias{collectPopIncidence}
\title{Collect population incidence estimates}
\usage{
collectPopIncidence(
  cdm,
  denominatorTable,
  outcomesTable,
  denominatorCohortIds = NULL,
  outcomeCohortIds = NULL,
  interval = "months",
  fullPeriodsRequired = TRUE,
  outcomeWashout = 0,
  repeatedEvents = FALSE,
  confidenceInterval = "poisson",
  minCellCount = 5,
  verbose = FALSE
)
}
\arguments{
\item{cdm}{CDMConnector CDM reference}

\item{denominatorTable}{Name of the table with the denominator cohorts}

\item{outcomesTable}{Name of the table with the outcome cohorts.
Can be "days", "weeks", "months", "quarters", or "years".}

\item{denominatorCohortIds}{Cohort ids of denominator populations}

\item{outcomeCohortIds}{Outcome cohort ids}

\item{interval}{Time intervals for incidence estimates}

\item{fullPeriodsRequired}{If full period is required}

\item{outcomeWashout}{Clean windows}

\item{repeatedEvents}{Repeated events}

\item{confidenceInterval}{Method for confidence intervals}

\item{minCellCount}{Minimum number of events to report- results
lower than this will be obscured. If NULL all results will be reported.}

\item{verbose}{Whether to report progress}
}
\description{
Collect population incidence estimates
}
