% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectPopPrevalence.R
\name{collectPopPrevalence}
\alias{collectPopPrevalence}
\title{Collect population prevalence estimates}
\usage{
collectPopPrevalence(
  cdm,
  denominatorTable,
  outcomesTable,
  denominatorCohortIds = NULL,
  outcomeCohortIds = NULL,
  type = "point",
  interval = "months",
  fullPeriodsRequired = TRUE,
  points = "start",
  minContribution = 0.5,
  confidenceInterval = "binomial",
  minCellCount = 5,
  verbose = FALSE
)
}
\arguments{
\item{cdm}{CDMConnector CDM reference object}

\item{denominatorTable}{denominatorTable}

\item{outcomesTable}{Name of the table with the outcome cohorts}

\item{denominatorCohortIds}{Cohort ids of denominator populations}

\item{outcomeCohortIds}{Outcome cohort ids}

\item{type}{type of prevalence, point or period}

\item{interval}{Time intervals for prevalence estimates}

\item{fullPeriodsRequired}{If full period is required}

\item{points}{where to compute the point prevalence}

\item{minContribution}{Minimum proportions that
individuals must have to contribute}

\item{confidenceInterval}{Method for confidence intervals}

\item{minCellCount}{Minimum number of events to report- results
lower than this will be obscured. If NULL all results will be reported.}

\item{verbose}{Whether to report progress}
}
\value{

}
\description{
Collect population prevalence estimates
}
