% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeIncidence.R
\name{computeIncidence}
\alias{computeIncidence}
\title{Collect population incidence estimates}
\usage{
computeIncidence(
  cdm,
  denominatorTable,
  outcomeTable,
  denominatorId = NULL,
  outcomeId = NULL,
  interval = "months",
  fullPeriods = TRUE,
  outcomeWashout = 0,
  repeatedEvents = FALSE,
  confidenceInterval = "poisson",
  minCellCount = 5,
  verbose = FALSE
)
}
\arguments{
\item{cdm}{CDMConnector CDM reference}

\item{denominatorTable}{Name of the table with the denominator cohorts}

\item{outcomeTable}{Name of the table with the outcome cohorts.
Can be "days", "weeks", "months", "quarters", or "years".}

\item{denominatorId}{Cohort ids of denominator populations}

\item{outcomeId}{Outcome cohort ids}

\item{interval}{Time intervals for incidence estimates}

\item{fullPeriods}{If full period is required}

\item{outcomeWashout}{Clean windows}

\item{repeatedEvents}{Repeated events}

\item{confidenceInterval}{Method for confidence intervals}

\item{minCellCount}{Minimum number of events to report- results
lower than this will be obscured. If NULL all results will be reported.}

\item{verbose}{Whether to report progress}
}
\description{
Collect population incidence estimates
}
