% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computePrevalence.R
\name{computePrevalence}
\alias{computePrevalence}
\title{Collect population prevalence estimates}
\usage{
computePrevalence(
  cdm,
  denominatorTable,
  outcomeTable,
  denominatorId = NULL,
  outcomeId = NULL,
  type = "point",
  interval = "months",
  fullPeriods = TRUE,
  points = "start",
  minContribution = 0.5,
  confidenceInterval = "binomial",
  minCellCount = 5,
  verbose = FALSE
)
}
\arguments{
\item{cdm}{CDMConnector CDM reference object}

\item{denominatorTable}{denominatorTable}

\item{outcomeTable}{Name of the table with the outcome cohorts}

\item{denominatorId}{Cohort ids of denominator populations}

\item{outcomeId}{Outcome cohort ids}

\item{type}{type of prevalence, point or period}

\item{interval}{Time intervals for prevalence estimates}

\item{fullPeriods}{If full period is required}

\item{points}{where to compute the point prevalence}

\item{minContribution}{Minimum proportions that
individuals must have to contribute}

\item{confidenceInterval}{Method for confidence intervals}

\item{minCellCount}{Minimum number of events to report- results
lower than this will be obscured. If NULL all results will be reported.}

\item{verbose}{Whether to report progress}
}
\description{
Collect population prevalence estimates
}
