% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_denominator_pop.R
\name{create_attrition_tibble}
\alias{create_attrition_tibble}
\title{Creates an attrition tibble}
\usage{
create_attrition_tibble(db_table, reason = NULL)
}
\arguments{
\item{db_table}{Database table that contains a person_id}

\item{reason}{the reason for the attrition}
}
\value{
a tibble
}
\description{
Creates an attrition tibble
}
