% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropStemTables.R
\name{dropStemTables}
\alias{dropStemTables}
\title{Drop stem tables}
\usage{
dropStemTables(cdm, computePermanentStem)
}
\arguments{
\item{cdm}{A CDM reference object}

\item{computePermanentStem}{The stem for which to drop tables. For example,
a stem of "study_results" would lead to dropping any table starting with this
(ie both "study_results" and "study_results_analysis_1" would be dropped if
they exist in the write schema)}
}
\description{
Drop stem tables
}
\examples{
\dontrun{
 cdm <- mockIncidencePrevalenceRef(sampleSize = 10000)
 cdm$dpop <- generateDenominatorCohortSet(
  cdm = cdm,
  computePermanent = TRUE,
  computePermanentStem = "result")
inc <- estimateIncidence(
  cdm = cdm,
  denominatorTable = "dpop",
  outcomeTable = "outcome",
  interval = "overall",
  computePermanent = TRUE,
  computePermanentStem = "result",
  returnParticipants = TRUE)
CDMConnector::listTables(attr(cdm, "dbcon"),
                          schema = attr(cdm, "write_schema")
)

dropStemTables(cdm, computePermanentStem = "result")
CDMConnector::listTables(attr(cdm, "dbcon"),
                         schema = attr(cdm, "write_schema")
}
}
