% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateRollingIncidence.R
\name{estimateRollingIncidence}
\alias{estimateRollingIncidence}
\title{Collect rolling population incidence estimates}
\usage{
estimateRollingIncidence(
  cdm,
  denominatorTable,
  outcomeTable,
  denominatorCohortId = NULL,
  outcomeCohortId = NULL,
  interval = "months",
  nIntervals = 12,
  outcomeWashout = Inf,
  repeatedEvents = FALSE,
  strata = list(),
  includeOverallStrata = TRUE,
  rateDenominator = 1e+05
)
}
\arguments{
\item{cdm}{A CDM reference object}

\item{denominatorTable}{A cohort table with a set of denominator cohorts
(for example, created using the \code{generateDenominatorCohortSet()}
function).}

\item{outcomeTable}{A cohort table in the cdm reference containing
a set of outcome cohorts.}

\item{denominatorCohortId}{The cohort definition ids or the cohort names of
the denominator cohorts of interest. If NULL all cohorts will be considered
in the analysis.}

\item{outcomeCohortId}{The cohort definition ids or the cohort names of the
outcome cohorts of interest. If NULL all cohorts will be considered in the
analysis.}

\item{interval}{Time intervals over which rolling incidence is estimated. Can
be "weeks", "months", "quarters", or "years". ISO weeks will
be used for weeks. Calendar months, quarters, or years can be used. If more
than one option is chosen then results will be estimated for each chosen interval.}

\item{nIntervals}{The total number of relative-time windows to estimate (e.g.,
12 means 12 windows of the specified interval).}

\item{outcomeWashout}{The number of days used for a 'washout' period
between the end of one outcome and an individual starting to contribute
time at risk. If Inf, no time can be contributed after an event has
occurred.}

\item{repeatedEvents}{TRUE/ FALSE. If TRUE, an individual will be able to
contribute multiple events during the study period (time while they are
present in an outcome cohort and any subsequent washout will be
excluded). If FALSE, an individual will only contribute time up to their
first event.}

\item{strata}{Variables added to the denominator cohort table for which to
stratify estimates.}

\item{includeOverallStrata}{Whether to include an overall result as well as
strata specific results (when strata has been specified).}

\item{rateDenominator}{The denominator to use for the incidence rate
calculation. Default is 100000.}
}
\value{
Rolling incidence estimates
}
\description{
Collect rolling population incidence estimates
}
\examples{
\donttest{
cdm <- mockIncidencePrevalence(sampleSize = 1000)
cdm <- generateDenominatorCohortSet(
  cdm = cdm, name = "denominator",
  cohortDateRange = c(as.Date("2008-01-01"), as.Date("2018-01-01"))
)
rolling_inc <- estimateRollingIncidence(
  cdm               = cdm,
  denominatorTable  = "denominator",
  outcomeTable      = "outcome",
  interval          = "months",
  nIntervals        = 12
)
}
}
