% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatherIncidencePrevalenceResults.R
\name{gatherIncidencePrevalenceResults}
\alias{gatherIncidencePrevalenceResults}
\title{Gather incidence and prevalence results}
\usage{
gatherIncidencePrevalenceResults(
  resultList,
  outcomeCohortId = NULL,
  outcomeCohortName = NULL,
  databaseName = NULL
)
}
\arguments{
\item{resultList}{List of incididence and prevalence results}

\item{outcomeCohortId}{Vector of cohort id for outcomes cohort}

\item{outcomeCohortName}{Vector of names for outcomes cohort corresponding to
the outcomeCohortId}

\item{databaseName}{A database name to add to to the results}
}
\value{
A list of up to two tibbles, one with prevalence results and one
with incidence results. Note, where multiple results sets of the same type
have been passed to the function, the analysisId variable will have been
updated
}
\description{
Gather incidence and prevalence results
}
