% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_mock_incidence_prevalence_db.R
\name{generate_mock_incidence_prevalence_db}
\alias{generate_mock_incidence_prevalence_db}
\title{Generate example database for IncidencePrevalence package}
\usage{
generate_mock_incidence_prevalence_db(
  person = NULL,
  observation_period = NULL,
  outcome = NULL
)
}
\arguments{
\item{person}{Person table.}

\item{observation_period}{observation_period table.}

\item{outcome}{outcome table.}
}
\value{
DBIConnection to duckdb database with mock data
}
\description{
Generate example database for IncidencePrevalence package
}
\examples{
\dontrun{
library(DBI)
library(duckdb)
library(tibble)
library(IncidencePrevalence)
person_example <- tibble(
  person_id = "1",
  gender_concept_id = "8507",
  year_of_birth = 2000,
  month_of_birth = 06,
  day_of_birth = 01
)
observation_period_example <- tibble(
  observation_period_id = "1",
  person_id = "1",
  observation_period_start_date = as.Date("2010-01-01"),
  observation_period_end_date = as.Date("2015-06-01")
)
db <- generate_mock_incidence_prevalence_db(
  person = person_example,
  observation_period = observation_period_example
)
}

}
