% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_confidence_intervals.R
\name{get_confidence_intervals}
\alias{get_confidence_intervals}
\title{Get confidence intervals}
\usage{
get_confidence_intervals(x, confidence_intervals)
}
\arguments{
\item{x}{x}

\item{confidence_intervals}{Confidence intervals type}
}
\description{
Get confidence intervals
}
