% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_denominator_pop.R
\name{get_denominator_pop}
\alias{get_denominator_pop}
\title{Identify the denominator populations}
\usage{
get_denominator_pop(
  cdm,
  start_date,
  end_date,
  min_age,
  max_age,
  days_prior_history,
  table_name_strata,
  strata_cohort_id,
  sample
)
}
\arguments{
\item{cdm}{CDMConnector CDM reference object}

\item{start_date}{Date indicating the start of the study period.}

\item{end_date}{Date indicating the end of the study period.}

\item{min_age}{Minimum ages for the cohort}

\item{max_age}{Maximum ages for the cohort}

\item{days_prior_history}{Days of prior history required to enter
the study cohort.}

\item{table_name_strata}{table_name_strata}

\item{strata_cohort_id}{strata_cohort_id}

\item{sample}{sample n}
}
\value{

}
\description{
Identify the denominator populations
}
