% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pop_incidence.R
\name{get_pop_incidence}
\alias{get_pop_incidence}
\title{Title}
\usage{
get_pop_incidence(
  db,
  results_schema_outcome,
  table_name_outcome,
  cohort_id_outcome = NULL,
  study_denominator_pop,
  cohort_id_denominator_pop = NULL,
  time_interval = c("Months"),
  prior_event_lookback = NULL,
  repetitive_events = FALSE,
  confidence_interval = "exact",
  verbose = FALSE
)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Title
}
