% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDenominatorPop.R
\name{getDenominatorPop}
\alias{getDenominatorPop}
\title{Identify the denominator populations}
\usage{
getDenominatorPop(
  cdm,
  startDate,
  endDate,
  minAge,
  maxAge,
  daysPriorHistory,
  strataTable,
  strataCohortId,
  sample
)
}
\arguments{
\item{cdm}{CDMConnector CDM reference object}

\item{startDate}{Date indicating the start of the study period.}

\item{endDate}{Date indicating the end of the study period.}

\item{minAge}{Minimum ages for the cohort}

\item{maxAge}{Maximum ages for the cohort}

\item{daysPriorHistory}{Days of prior history required to enter
the study cohort.}

\item{strataTable}{strataTable}

\item{strataCohortId}{strataCohortId}

\item{sample}{sample n}
}
\value{

}
\description{
Identify the denominator populations
}
