% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPopIncidence.R
\name{getIncidence}
\alias{getIncidence}
\title{Get population incidence estimates}
\usage{
getIncidence(
  cdm,
  denominatorTable,
  outcomeTable,
  denominatorId,
  outcomeId,
  interval,
  fullPeriods,
  outcomeWashout,
  repeatedEvents,
  returnAnalysisSettings,
  returnAttrition,
  returnPopulation,
  verbose
)
}
\arguments{
\item{cdm}{CDMConnector CDM reference object}

\item{denominatorTable}{denominatorTable}

\item{outcomeTable}{outcomeTable}

\item{denominatorId}{denominatorId}

\item{outcomeId}{outcomeId}

\item{interval}{interval}

\item{fullPeriods}{fullPeriods}

\item{outcomeWashout}{outcomeWashout}

\item{repeatedEvents}{repeatedEvents}

\item{returnAnalysisSettings}{Whether to return analysis settings}

\item{returnAttrition}{Whether to return attrition}

\item{returnPopulation}{Whether to return population}

\item{verbose}{verbose}
}
\description{
Get population incidence estimates
}
