% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidencePrevalence-package.R
\docType{package}
\name{IncidencePrevalence-package}
\alias{IncidencePrevalence}
\alias{IncidencePrevalence-package}
\title{IncidencePrevalence: Population Incidence and Prevalence using the OMOP Common Data Model}
\description{
Calculate population incidence and prevalence using data mapped to the OMOP common data model.
}
\author{
\strong{Maintainer}: Edward Burn \email{edward.burn@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0002-9286-1128}{ORCID})

Authors:
\itemize{
  \item Berta Raventos \email{braventos@idiapjgol.info} (\href{https://orcid.org/0000-0002-4668-2970}{ORCID})
  \item Marti Catala \email{marti.catalasabate@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0003-3308-9905}{ORCID})
  \item Mike Du \email{mike.du@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0002-9517-8834}{ORCID})
}

}
\keyword{internal}
