% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IncidencePrevalenceBenchmarkResults}
\alias{IncidencePrevalenceBenchmarkResults}
\title{Benchmarking results}
\format{
A list of results from benchmarking
}
\usage{
IncidencePrevalenceBenchmarkResults
}
\description{
Benchmarking results
}
\keyword{datasets}
