% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{incidenceSet}
\alias{incidenceSet}
\title{Settings associated with an incidence analysis}
\usage{
incidenceSet(result)
}
\arguments{
\item{result}{Result for which to get settings}
}
\value{
tibble with settings used when estimating incidence
}
\description{
Settings associated with an incidence analysis
}
