% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockIncidencePrevalenceRef.R
\name{mockIncidencePrevalenceRef}
\alias{mockIncidencePrevalenceRef}
\title{Generate example database for IncidencePrevalence package}
\usage{
mockIncidencePrevalenceRef(
  personTable = NULL,
  observationPeriodTable = NULL,
  strataTable = NULL,
  outcomeTable = NULL,
  sampleSize = 1,
  outPre = 1,
  seed = 444,
  ageBeta = NULL,
  genderBeta = NULL,
  intercept = NULL,
  earliestDateOfBirth = NULL,
  latestDateOfBirth = NULL,
  earliestObservationStartDate = NULL,
  latestObservationStartDate = NULL,
  minDaysToObservationEnd = NULL,
  maxDaysToObservationEnd = NULL,
  minOutcomeDays = 1,
  maxOutcomeDays = 10,
  maxOutcomes = 1
)
}
\arguments{
\item{personTable}{Person table.}

\item{observationPeriodTable}{observation_period table.}

\item{strataTable}{strata table.}

\item{outcomeTable}{outcome table.}

\item{sampleSize}{number of unique patient}

\item{outPre}{fraction of patient with an event}

\item{seed}{seed for simulating the data set use same seed to get same data set}

\item{ageBeta}{the beta for the standardized age in logistics regression outcome model}

\item{genderBeta}{the beta for the gender flag in logistics regression outcome model}

\item{intercept}{the beta for the intercept in the logistics regression outcome model}

\item{earliestDateOfBirth}{the earliest date of birth of patient in person table format "dd-mm-yyyy"}

\item{latestDateOfBirth}{the latest date of birth for patient in person table format "dd-mm-yyyy"}

\item{earliestObservationStartDate}{the earliest observation start date for patient format "dd-mm-yyyy"}

\item{latestObservationStartDate}{the latest observation start date for patient format "dd-mm-yyyy"}

\item{minDaysToObservationEnd}{the minimum number of days of the observational integer}

\item{maxDaysToObservationEnd}{the maximum number of days of the observation period integer}

\item{minOutcomeDays}{the minimum number of days of the outcome period default set to 1}

\item{maxOutcomeDays}{the maximum number of days of the outcome period default set to 10}

\item{maxOutcomes}{the maximum possible number of outcomes per person can have default set to 1}
}
\value{
CDMConnector CDM reference object to duckdb database with mock data
}
\description{
Generate example database for IncidencePrevalence package
}
\examples{
\dontrun{
library(DBI)
library(duckdb)
library(tibble)
library(IncidencePrevalence)
person_example <- tibble(
  person_id = "1",
  gender_concept_id = "8507",
  year_of_birth = 2000,
  month_of_birth = 06,
  day_of_birth = 01
)
observation_period_example <- tibble(
  observation_period_id = "1",
  person_id = "1",
  observation_period_start_date = as.Date("2010-01-01"),
  observation_period_end_date = as.Date("2015-06-01")
)
db <- mockIncidencePrevalenceRef(
  person = person_example,
  observation_period = observation_period_example
)
}

}
