% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obscure_counts.R
\name{obscure_counts}
\alias{obscure_counts}
\title{Obscure the small number of counts}
\usage{
obscure_counts(x, minimum_cell_count = 5, substitute = NA)
}
\arguments{
\item{x}{x}

\item{minimum_cell_count}{minimum_cell_count}

\item{substitute}{substitute}
}
\description{
Obscure the small number of counts
}
