% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obscureCounts.R
\name{obscureCounts}
\alias{obscureCounts}
\title{Obscure the small number of counts}
\usage{
obscureCounts(x, minCellCount = 5, substitute = NA)
}
\arguments{
\item{x}{x}

\item{minCellCount}{minCellCount}

\item{substitute}{substitute}
}
\description{
Obscure the small number of counts
}
