% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{optionsTableIncidence}
\alias{optionsTableIncidence}
\title{Additional arguments for the functions tableIncidence.}
\usage{
optionsTableIncidence()
}
\value{
The default .options named list.
}
\description{
It provides a list of allowed inputs for .option argument in tableIncidence,
and their given default values.
}
\examples{
{
optionsTableIncidence()
}

}
