% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{optionsTableIncidencePrevalence}
\alias{optionsTableIncidencePrevalence}
\title{Additional arguments for the functions tablePointPrevalence,
tablePeriodPrevalence, and tableIncidence.}
\usage{
optionsTableIncidencePrevalence()
}
\value{
The default .options named list.
}
\description{
It provides a list of allowed inputs for .option argument in the table
functions of the package IncidencePrevalence, and their given default values.
}
\examples{
{
optionsTableIncidencePrevalence()
}

}
