% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{optionsTablePrevalence}
\alias{optionsTablePrevalence}
\title{Additional arguments for the functions tablePrevalence.}
\usage{
optionsTablePrevalence()
}
\value{
The default .options named list.
}
\description{
It provides a list of allowed inputs for .option argument in tablePrevalence,
and their given default values.
}
\examples{
{
  optionsTablePrevalence()
}

}
