% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{participants}
\alias{participants}
\title{Participants contributing to an analysis}
\usage{
participants(x)
}
\arguments{
\item{x}{Result object}
}
\value{
References to tables with the study participants contributing to
a given analysis
}
\description{
Participants contributing to an analysis
}
