% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotRollingIncidence}
\alias{plotRollingIncidence}
\title{Plot rolling incidence results}
\usage{
plotRollingIncidence(
  result,
  x = "window_start_days",
  y = "incidence_100000_pys",
  line = TRUE,
  point = TRUE,
  ribbon = FALSE,
  ymin = "incidence_100000_pys_95CI_lower",
  ymax = "incidence_100000_pys_95CI_upper",
  facet = NULL,
  colour = NULL
)
}
\arguments{
\item{result}{Rolling incidence results}

\item{x}{Variable to plot on x axis}

\item{y}{Variable to plot on y axis.}

\item{line}{Whether to plot a line using \code{geom_line}}

\item{point}{Whether to plot points using \code{geom_point}}

\item{ribbon}{Whether to plot a ribbon using \code{geom_ribbon}}

\item{ymin}{Lower limit of error bars, if provided is plot using \code{geom_errorbar}}

\item{ymax}{Upper limit of error bars, if provided is plot using \code{geom_errorbar}}

\item{facet}{Variables to use for facets. To see available variables for
facetting use the function \code{availableRollingIncidenceGrouping()}.}

\item{colour}{Variables to use for colours. To see available variables for
colouring use the function \code{availableRollingIncidenceGrouping()}.}
}
\value{
A ggplot with the rolling incidence results plotted
}
\description{
Plot rolling incidence results
}
