% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.IncidencePrevalenceDenominator}
\alias{print.IncidencePrevalenceDenominator}
\title{Print a CDM reference object}
\usage{
\method{print}{IncidencePrevalenceDenominator}(x, ...)
}
\arguments{
\item{x}{A cdm_reference object}

\item{...}{Included for compatibility with generic. Not used.}
}
\value{
Invisibly returns the input
}
\description{
Print a CDM reference object
}
