% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recordAttrition.R
\name{record_attrition}
\alias{record_attrition}
\title{Record attrition in tibble}
\usage{
record_attrition(
  table,
  id = "person_id",
  reason = NULL,
  existing_attrition = NULL
)
}
\arguments{
\item{table}{table that contains a person_id}

\item{id}{id variable}

\item{reason}{the reason for the attrition}

\item{existing_attrition}{previous attrition to append results}
}
\value{
a tibble
}
\description{
Record attrition in tibble
}
