% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recordAttrition.R
\name{recordAttrition}
\alias{recordAttrition}
\title{Record attrition in tibble}
\usage{
recordAttrition(
  table,
  id = "person_id",
  reason = NULL,
  existingAttrition = NULL
)
}
\arguments{
\item{table}{table that contains a person_id}

\item{id}{id variable}

\item{reason}{the reason for the attrition}

\item{existingAttrition}{previous attrition to append results}
}
\value{
a tibble
}
\description{
Record attrition in tibble
}
