% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-omopgenerics.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{cohortCount}
\alias{settings}
\alias{attrition}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{omopgenerics}{\code{\link[omopgenerics]{attrition}}, \code{\link[omopgenerics]{cohortCount}}, \code{\link[omopgenerics]{settings}}}
}}

