% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sqlTrace}
\alias{sqlTrace}
\title{SQL trace for a cohort}
\usage{
sqlTrace(x)
}
\arguments{
\item{x}{Cohort set for which to get associated SQL}
}
\value{
SQL used to generate the cohort set
}
\description{
SQL trace for a cohort
}
