% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tableIncidence}
\alias{tableIncidence}
\title{Format a point_prevalence object into a visual table.}
\usage{
tableIncidence(result, type = "gt", .options = list())
}
\arguments{
\item{result}{A summarised_result object with results from
estimateIncidence().}

\item{type}{Type of desired formatted table, possibilities: "gt",
"flextable", "tibble".}

\item{.options}{Named list with additional formatting options.
IncidencePrevalence::optionsTableIncidence() shows allowed
arguments and their default values.}
}
\value{
A table with a formatted version of incidence results.
}
\description{
Format a point_prevalence object into a visual table.
}
\examples{
\donttest{
library(IncidencePrevalence)

cdm <- mockIncidencePrevalenceRef()

cdm <- generateDenominatorCohortSet(cdm = cdm, name = "denominator")

inc <- estimateIncidence(
  cdm = cdm,
  denominatorTable = "denominator",
  outcomeTable = "outcome"
)

tableIncidence(inc)

CDMConnector::cdmDisconnect(cdm = cdm)
}

}
