% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tablePrevalence}
\alias{tablePrevalence}
\title{Table of prevalence results}
\usage{
tablePrevalence(result, type = "gt", .options = list())
}
\arguments{
\item{result}{Prevalence results}

\item{type}{Type of table. Can be "gt", "flextable", or "tibble".}

\item{.options}{Table options to apply}
}
\value{
Table of prevalence results
}
\description{
Table of prevalence results
}
