% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tablePrevalenceAttrition}
\alias{tablePrevalenceAttrition}
\title{Table of prevalence attrition results}
\usage{
tablePrevalenceAttrition(
  result,
  type = "gt",
  header = "variable_name",
  groupColumn = c("cdm_name", "variable_level"),
  settingsColumn = omopgenerics::settingsColumns(result),
  hide = "estimate_name",
  settingsColumns = lifecycle::deprecated()
)
}
\arguments{
\item{result}{A summarised_result object. Output of
summariseCohortAttrition().}

\item{type}{Type of table. Check supported types with
\code{visOmopResults::tableType()}.}

\item{header}{Columns to use as header. See options with
\code{colnames(omopgenerics::splitAll(result))}. Variables in \code{settingsColumn}
are also allowed}

\item{groupColumn}{Variables to use as group labels. Allowed columns are the
same as in \code{header}}

\item{settingsColumn}{Variables from the settings atribute to dispaly in
the table}

\item{hide}{Table columns to exclude, options are the ones described in
\code{header}}

\item{settingsColumns}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
}
\value{
A visual table.
}
\description{
Table of prevalence attrition results
}
