% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeResultToDisk.R
\name{writeResultToDisk}
\alias{writeResultToDisk}
\title{Export IncidencePrevalence results}
\usage{
writeResultToDisk(result, zipName, outputFolder)
}
\arguments{
\item{result}{IncidencePrevalence results from gatherResults()}

\item{zipName}{name to give zip folder}

\item{outputFolder}{directory to save zip folder containing results as a set
of CSV files}
}
\value{
zip folder of results saved in the outputFolder
}
\description{
Export IncidencePrevalence results
}
