% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{attrition}
\alias{attrition}
\title{Cohort attrition}
\usage{
attrition(result, analysisId = NULL)
}
\arguments{
\item{result}{Result for which to get attrition}

\item{analysisId}{ID of a specific analysis to return attrition for}
}
\value{
tibble with counts and reasons for attrition.
}
\description{
Cohort attrition
}
