% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarkIncidencePrevalence.R
\name{benchmarkIncidencePrevalence}
\alias{benchmarkIncidencePrevalence}
\title{Run benchmark of incidence and prevalence analyses}
\usage{
benchmarkIncidencePrevalence(
  cdm,
  sample = NULL,
  nOutcomes = 1,
  prevOutcomes = c(25),
  analysisType = 1,
  outputFolder = NULL,
  verbose = FALSE
)
}
\arguments{
\item{cdm}{A CDM reference object}

\item{sample}{An integer for which to take a random sample when generating
the denominator cohort}

\item{nOutcomes}{An integer specifying the number of outcomes to create in the denominator cohort.
Default is 1.}

\item{prevOutcomes}{An array of integers for the prevalence of the outcomes in the population (in \%).
Default is 25\%.}

\item{analysisType}{An integer for the analysis the user wants to perform:
1 for all, 2 for no incidence, 3 for no point prevalence, 4 for no period prevalence,
5 for only incidence, 6 for only point prevalence, 7 for only period prevalence.
Default is 1 (all).}

\item{outputFolder}{Folder to save results as CSV}

\item{verbose}{Either TRUE or FALSE. If TRUE, progress will be reported.}
}
\value{
a tibble with time taken for different analyses
}
\description{
Run benchmark of incidence and prevalence analyses
}
\examples{
\dontrun{
con <- DBI::dbConnect(duckdb::duckdb(), CDMConnector::eunomia_dir())
cdm <- CDMConnector::cdm_from_con(
  con = con,
  cdm_schema = "main"
)

timings <- IncidencePrevalence::benchmarkIncidencePrevalence(cdm)
}
}
