% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarkIncidencePrevalence.R
\name{benchmarkIncidencePrevalence}
\alias{benchmarkIncidencePrevalence}
\title{Run benchmark of incidence and prevalence analyses}
\usage{
benchmarkIncidencePrevalence(
  cdm,
  type = "simple",
  denominatorSampleSize = NULL,
  outcomeTableBench = NULL,
  estimationInterval = "years",
  verbose = FALSE
)
}
\arguments{
\item{cdm}{A CDM reference object}

\item{type}{Whether to run a "simple", "typical", or "complex" analysis}

\item{denominatorSampleSize}{Number for a random sample when generating
denominator cohorts}

\item{outcomeTableBench}{A cohort table in the cdm reference containing
a set of outcome cohorts. All cohorts in the table will be used.}

\item{estimationInterval}{Interval used for estimating incidence and
prevalence. Can be "weeks", "months", "quarters", or "years"}

\item{verbose}{Either TRUE or FALSE. If TRUE, progress will be reported.}
}
\value{
a tibble with time taken for different analyses
}
\description{
Run benchmark of incidence and prevalence analyses
}
