% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_denominator_pops.R
\name{collect_denominator_pops}
\alias{collect_denominator_pops}
\title{Identify a set of denominator populations}
\usage{
collect_denominator_pops(
  db,
  cdm_database_schema,
  study_start_date = NULL,
  study_end_date = NULL,
  study_age_stratas = NULL,
  study_sex_stratas = "Both",
  study_days_prior_history = 0,
  verbose = FALSE
)
}
\arguments{
\item{db}{Database connection via DBI::dbConnect()}

\item{cdm_database_schema}{Name of the schema which contains the
omop cdm person and observation_period tables}

\item{study_start_date}{Date indicating the start of the study
period. If NULL,
the earliest observation_start_date in the observation_period table
will be used.}

\item{study_end_date}{Date indicating the end of the study
period. If NULL,
the latest observation_end_date in the observation_period table
will be used.}

\item{study_age_stratas}{List of age groups}

\item{study_sex_stratas}{Sex of the cohorts}

\item{study_days_prior_history}{Days of prior history required to enter
the study cohort.}

\item{verbose}{Either TRUE or FALSE.
If TRUE, progress will be reported.}
}
\description{
Identify a set of denominator populations
}
