% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_denominator_pops.R
\name{collect_denominator_pops}
\alias{collect_denominator_pops}
\title{Identify a set of denominator populations}
\usage{
collect_denominator_pops(
  cdm,
  study_start_date = NULL,
  study_end_date = NULL,
  study_age_stratas = list(c(0, 150)),
  study_sex_stratas = "Both",
  study_days_prior_history = 0,
  combine_subgroups = TRUE,
  table_name_strata = NULL,
  strata_cohort_id = NULL,
  sample = NULL,
  verbose = FALSE
)
}
\arguments{
\item{cdm}{CDMConnector CDM reference object}

\item{study_start_date}{Date indicating the start of the study
period. If NULL,
the earliest observation_start_date in the observation_period table
will be used.}

\item{study_end_date}{Date indicating the end of the study
period. If NULL,
the latest observation_end_date in the observation_period table
will be used.}

\item{study_age_stratas}{List of age groups}

\item{study_sex_stratas}{Sex of the cohorts}

\item{study_days_prior_history}{Days of prior history required to enter
the study cohort.}

\item{combine_subgroups}{If TRUE all combinations of age, sex, and
days prior history will be created. If FALSE age, sex, and
days prior history will be considered separately.}

\item{table_name_strata}{table_name_strata}

\item{strata_cohort_id}{strata_cohort_id}

\item{sample}{sample n}

\item{verbose}{Either TRUE or FALSE.
If TRUE, progress will be reported.}
}
\description{
\code{collect_denominator_pops()} creates a set of denominator cohorts
}
\examples{
\dontrun{
db <- DBI::dbConnect(" Your database connection here ")
cdm <- CDMConnector::cdm_from_con(con = db,
                                  cdm_schema = "cdm schema name")
dpop <- collect_denominator_pops(
 cdm = cdm,
 study_start_date = as.Date("2008-01-01"),
 study_end_date = as.Date("2018-01-01")
)
}
}
