% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_pop_prevalence.R
\name{collect_pop_prevalence}
\alias{collect_pop_prevalence}
\title{Collect population prevalence estimates}
\usage{
collect_pop_prevalence(
  db,
  results_schema_outcomes,
  table_name_outcomes,
  cohort_ids_outcomes,
  study_denominator_pop,
  cohort_ids_denominator_pops,
  periods = "Point",
  time_intervals = "Months",
  minimum_representative_proportions = 0.5,
  confidence_intervals = "exact",
  minimum_event_count = 5,
  verbose = FALSE
)
}
\arguments{
\item{db}{Database connection via DBI::dbConnect()}

\item{results_schema_outcomes}{Name of the schema which contains
the outcome table}

\item{table_name_outcomes}{Name of the table with the outcome cohorts}

\item{cohort_ids_outcomes}{Outcome cohort ids}

\item{study_denominator_pop}{Tibble with denominator populations}

\item{cohort_ids_denominator_pops}{Cohort ids of denominator populations}

\item{periods}{Periods to compute the prevalence}

\item{time_intervals}{Time intervals for incidence estimates}

\item{minimum_representative_proportions}{Minimum proportions that
individuals must have to contribute}

\item{confidence_intervals}{Method for confidence intervals}

\item{minimum_event_count}{Minimum number of events to report- results
lower than this will be obscured. If NULL all results will be reported.}

\item{verbose}{Whether to report progress}
}
\value{

}
\description{
Collect population prevalence estimates
}
