% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateIncidence.R
\name{estimateIncidence}
\alias{estimateIncidence}
\title{Collect population incidence estimates}
\usage{
estimateIncidence(
  cdm,
  denominatorTable,
  outcomeTable,
  denominatorCohortId = NULL,
  outcomeCohortId = NULL,
  interval = "months",
  fullPeriods = TRUE,
  outcomeWashout = 0,
  repeatedEvents = FALSE,
  confidenceInterval = "poisson",
  returnAnalysisCohort = TRUE,
  minCellCount = 5,
  verbose = FALSE
)
}
\arguments{
\item{cdm}{CDMConnector CDM reference}

\item{denominatorTable}{Name of the table with the denominator cohorts}

\item{outcomeTable}{Name of the table with the outcome cohorts.
Can be "weeks", "months", "quarters", or "years".}

\item{denominatorCohortId}{Cohort ids of denominator populations}

\item{outcomeCohortId}{Outcome cohort ids}

\item{interval}{Time intervals for incidence estimates}

\item{fullPeriods}{If full period is required}

\item{outcomeWashout}{Clean windows}

\item{repeatedEvents}{Repeated events}

\item{confidenceInterval}{Method for confidence intervals}

\item{returnAnalysisCohort}{Whether to return population}

\item{minCellCount}{Minimum number of events to report- results
lower than this will be obscured. If NULL all results will be reported.}

\item{verbose}{Whether to report progress}
}
\description{
Collect population incidence estimates
}
