% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatePrevalence.R
\name{estimatePointPrevalence}
\alias{estimatePointPrevalence}
\title{Estimate point prevalence}
\usage{
estimatePointPrevalence(
  cdm,
  denominatorTable,
  outcomeTable,
  denominatorCohortId = NULL,
  outcomeCohortId = NULL,
  outcomeLookbackDays = 0,
  interval = "months",
  completeDatabaseIntervals = TRUE,
  points = "start",
  confidenceInterval = "binomial",
  minCellCount = 5,
  verbose = FALSE
)
}
\arguments{
\item{cdm}{A CDM reference object}

\item{denominatorTable}{A cohort table with a set of denominator cohorts
(for example, created using the \code{generateDenominatorCohortSet()}
function).}

\item{outcomeTable}{A cohort table in the cdm reference containing
a set of outcome cohorts.}

\item{denominatorCohortId}{The cohort definition ids of the denominator
cohorts of interest. If NULL all cohorts will be considered in the
analysis.}

\item{outcomeCohortId}{The cohort definition ids of the outcome
cohorts of interest. If NULL all cohorts will be considered in the
analysis.}

\item{outcomeLookbackDays}{Days lookback when considering an outcome
as prevalent. If NULL any prior outcome will be considered as prevalent. If
0, only ongoing outcomes will be considered as prevalent.}

\item{interval}{Time intervals over which period prevalence is estimated. Can
be "weeks", "months", "quarters", "years", or "overall". ISO weeks will
be used for weeks. Calendar months, quarters, or years can be used as
the period. If more than one option is chosen then results will
be estimated for each chosen interval.}

\item{completeDatabaseIntervals}{TRUE/ FALSE. Where TRUE, incidence will
only be estimated for those intervals where the database
captures all the interval (based on the earliest and latest observation
period start dates, respectively).}

\item{points}{where to compute the point prevalence}

\item{confidenceInterval}{Method for confidence intervals}

\item{minCellCount}{Minimum number of events to report- results
lower than this will be obscured. If NULL all results will be reported.}

\item{verbose}{Whether to report progress}
}
\value{
Point prevalence estimates
}
\description{
Estimate point prevalence
}
\examples{
\dontrun{
db <- DBI::dbConnect(" Your database connection here ")
cdm <- CDMConnector::cdm_from_con(
  con = db,
  cdm_schema = "cdm schema name"
)
dpop <- generateDenominatorCohortSet(
  cdm = cdm,
  startDate = as.Date("2008-01-01"),
  endDate = as.Date("2018-01-01")
)
cdm$denominator <- dpop$denominator_population
estimatePointPrevalence(
  cdm = cdm,
  denominatorTable = "denominator",
  outcomeTable = "outcome",
  interval = "months"
)
}
}
