% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatePrevalence.R
\name{estimatePointPrevalence}
\alias{estimatePointPrevalence}
\title{Estimate point prevalence}
\usage{
estimatePointPrevalence(
  cdm,
  denominatorTable,
  outcomeTable,
  denominatorCohortId = NULL,
  outcomeCohortId = NULL,
  outcomeLookbackDays = 0,
  interval = "months",
  fullPeriods = TRUE,
  fullContribution = FALSE,
  points = "start",
  confidenceInterval = "binomial",
  minCellCount = 5,
  verbose = FALSE
)
}
\arguments{
\item{cdm}{CDMConnector CDM reference object}

\item{denominatorTable}{denominatorTable}

\item{outcomeTable}{Name of the table with the outcome cohorts}

\item{denominatorCohortId}{Cohort ids of denominator populations}

\item{outcomeCohortId}{Outcome cohort ids}

\item{outcomeLookbackDays}{Days lookback when considering an outcome as prevalent. If NULL any prior outcome will be considered as prevalent.}

\item{interval}{Time intervals for prevalence estimates}

\item{fullPeriods}{If full period is required}

\item{fullContribution}{If complete periods are required for
individuals to contribute}

\item{points}{where to compute the point prevalence}

\item{confidenceInterval}{Method for confidence intervals}

\item{minCellCount}{Minimum number of events to report- results
lower than this will be obscured. If NULL all results will be reported.}

\item{verbose}{Whether to report progress}
}
\description{
Estimate point prevalence
}
