% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_mock_incidence_prevalence_db.R
\name{generate_mock_incidence_prevalence_db}
\alias{generate_mock_incidence_prevalence_db}
\title{Generate example database for IncidencePrevalence package}
\usage{
generate_mock_incidence_prevalence_db(
  person = NULL,
  observation_period = NULL,
  outcome = NULL,
  sample_size = 1,
  out_pre = 1,
  seed = 444,
  age_beta = NULL,
  gender_beta = NULL,
  intercept = NULL,
  earliest_date_of_birth = NULL,
  latest_date_of_birth = NULL,
  earliest_observation_start_date = NULL,
  latest_observation_start_date = NULL,
  min_days_to_observation_end = NULL,
  max_days_to_observation_end = NULL,
  min_outcome_days = 1,
  max_outcome_days = 10,
  max_outcomes_per_person = 1
)
}
\arguments{
\item{person}{Person table.}

\item{observation_period}{observation_period table.}

\item{outcome}{outcome table.}

\item{sample_size}{number of unique patient}

\item{out_pre}{fraction of patient with an event}

\item{seed}{seed for simulating the data set use same seed to get same data set}

\item{age_beta}{the beta for the standardized age in logistics regression outcome model}

\item{gender_beta}{the beta for the gender flag in logistics regression outcome model}

\item{intercept}{the beta for the intercept in the logistics regression outcome model}

\item{earliest_date_of_birth}{the earliest date of birth of patient in person table format "dd-mm-yyyy"}

\item{latest_date_of_birth}{the latest date of birth for patient in person table format "dd-mm-yyyy"}

\item{earliest_observation_start_date}{the earliest observation start date for patient format "dd-mm-yyyy"}

\item{latest_observation_start_date}{the latest observation start date for patient format "dd-mm-yyyy"}

\item{min_days_to_observation_end}{the minimum number of days of the observational integer}

\item{max_days_to_observation_end}{the maximum number of days of the observation period integer}

\item{min_outcome_days}{the minimum number of days of the outcome period default set to 1}

\item{max_outcome_days}{the maximum number of days of the outcome period default set to 10}

\item{max_outcomes_per_person}{the maximum possible number of outcomes per person can have default set to 1}
}
\value{
DBIConnection to duckdb database with mock data
}
\description{
Generate example database for IncidencePrevalence package
}
\examples{
\dontrun{
library(DBI)
library(duckdb)
library(tibble)
library(IncidencePrevalence)
person_example <- tibble(
  person_id = "1",
  gender_concept_id = "8507",
  year_of_birth = 2000,
  month_of_birth = 06,
  day_of_birth = 01
)
observation_period_example <- tibble(
  observation_period_id = "1",
  person_id = "1",
  observation_period_start_date = as.Date("2010-01-01"),
  observation_period_end_date = as.Date("2015-06-01")
)
db <- generate_mock_incidence_prevalence_db(
  person = person_example,
  observation_period = observation_period_example
)
}

}
