% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDenominatorCohortSet.R
\name{generateDenominatorCohortSet}
\alias{generateDenominatorCohortSet}
\title{Identify a set of denominator populations}
\usage{
generateDenominatorCohortSet(
  cdm,
  startDate = NULL,
  endDate = NULL,
  ageGroups = list(c(0, 150)),
  sex = "Both",
  daysPriorHistory = 0,
  strataTable = NULL,
  strataCohortId = NULL,
  sample = NULL,
  verbose = FALSE
)
}
\arguments{
\item{cdm}{CDMConnector CDM reference object}

\item{startDate}{Date indicating the start of the study
period. If NULL,
the earliest observation_start_date in the observation_period table
will be used.}

\item{endDate}{Date indicating the end of the study
period. If NULL,
the latest observation_end_date in the observation_period table
will be used.}

\item{ageGroups}{List of age groups}

\item{sex}{Sex of the cohorts}

\item{daysPriorHistory}{Days of prior history required to enter
the study cohort.}

\item{strataTable}{strataTable}

\item{strataCohortId}{strata cohort definition id}

\item{sample}{sample n}

\item{verbose}{Either TRUE or FALSE.
If TRUE, progress will be reported.}
}
\description{
\code{generateDenominatorCohortSet()} creates a set of denominator cohorts
}
\examples{
\dontrun{
con <- DBI::dbConnect(duckdb::duckdb(), CDMConnector::eunomia_dir())
cdm <- CDMConnector::cdm_from_con(
  con = con,
  cdm_schema = "main"
)

dpop <- generateDenominatorCohortSet(
  cdm = cdm,
  startDate = as.Date("2008-01-01"),
  endDate = as.Date("2018-01-01")
)
}
}
