% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_denominator_pop.R
\name{get_denominator_pop}
\alias{get_denominator_pop}
\title{Identify a denominator population}
\usage{
get_denominator_pop(
  db,
  cdm_database_schema,
  start_date = NULL,
  end_date = NULL,
  min_age = NULL,
  max_age = NULL,
  sex = "Both",
  days_prior_history = NULL,
  verbose = FALSE
)
}
\arguments{
\item{db}{Database connection via DBI::dbConnect()}

\item{cdm_database_schema}{Name of the schema which contains the
omop cdm person and observation_period tables}

\item{start_date}{Date indicating the start of the study period. If NULL,
the earliest observation_start_date in the observation_period table
will be used.}

\item{end_date}{Date indicating the end of the study period. If NULL,
the latest observation_end_date in the observation_period table
will be used.}

\item{min_age}{Minimum age for the cohort}

\item{max_age}{Maximum age for the cohort.}

\item{sex}{Sex of the cohort}

\item{days_prior_history}{Days of prior history required to enter
the study cohort.}

\item{verbose}{Either TRUE or FALSE.
If TRUE, progress will be reported.}
}
\value{

}
\description{
Identify a denominator population
}
