% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_denominator_pop.R
\name{get_denominator_pop}
\alias{get_denominator_pop}
\title{Identify a single denominator population}
\usage{
get_denominator_pop(
  cdm_ref,
  start_date,
  end_date,
  min_age,
  max_age,
  days_prior_history,
  table_name_strata,
  strata_cohort_id
)
}
\arguments{
\item{cdm_ref}{CDMConnector CDM reference object}

\item{start_date}{Date indicating the start of the study period. If NULL,
the earliest observation_start_date in the observation_period table
will be used.}

\item{end_date}{Date indicating the end of the study period. If NULL,
the latest observation_end_date in the observation_period table
will be used.}

\item{min_age}{Minimum age for the cohort}

\item{max_age}{Maximum age for the cohort.}

\item{days_prior_history}{Days of prior history required to enter
the study cohort.}

\item{table_name_strata}{table_name_strata}

\item{strata_cohort_id}{strata_cohort_id}
}
\value{

}
\description{
Identify a single denominator population
}
