% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pop_incidence.R
\name{get_pop_incidence}
\alias{get_pop_incidence}
\title{Get population incidence estimates}
\usage{
get_pop_incidence(
  cdm_ref,
  table_name_denominator,
  cohort_id_denominator_pop,
  table_name_outcome,
  cohort_id_outcome,
  time_interval,
  outcome_washout_window,
  repetitive_events,
  verbose
)
}
\arguments{
\item{cdm_ref}{CDMConnector CDM reference object}

\item{table_name_denominator}{table_name_denominator}

\item{cohort_id_denominator_pop}{cohort_id_denominator_pop}

\item{table_name_outcome}{table_name_outcome}

\item{cohort_id_outcome}{cohort_id_outcome}

\item{time_interval}{time_interval}

\item{outcome_washout_window}{outcome_washout_window}

\item{repetitive_events}{repetitive_events}

\item{verbose}{verbose}
}
\value{

}
\description{
Get population incidence estimates
}
