% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPopIncidence.R
\name{get_pop_incidence}
\alias{get_pop_incidence}
\title{Get population incidence estimates}
\usage{
get_pop_incidence(
  cdm,
  denominatorTable,
  outcomesTable,
  denominatorCohortIds,
  outcomeCohortIds,
  interval,
  fullPeriodsRequired,
  outcomeWashout,
  repeatedEvents,
  verbose
)
}
\arguments{
\item{cdm}{CDMConnector CDM reference object}

\item{denominatorTable}{denominatorTable}

\item{outcomesTable}{outcomesTable}

\item{denominatorCohortIds}{denominatorCohortIds}

\item{outcomeCohortIds}{outcomeCohortIds}

\item{interval}{interval}

\item{fullPeriodsRequired}{fullPeriodsRequired}

\item{outcomeWashout}{outcomeWashout}

\item{repeatedEvents}{repeatedEvents}

\item{verbose}{verbose}
}
\value{

}
\description{
Get population incidence estimates
}
