% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pop_incidence.R
\name{get_pop_incidence}
\alias{get_pop_incidence}
\title{Get population incidence estimates}
\usage{
get_pop_incidence(
  db,
  results_schema_outcome,
  table_name_outcome,
  cohort_id_outcome = NULL,
  study_denominator_pop,
  cohort_id_denominator_pop = NULL,
  time_interval = c("months"),
  outcome_washout_window = NULL,
  repetitive_events = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{db}{db}

\item{results_schema_outcome}{results_schema_outcome}

\item{table_name_outcome}{table_name_outcome}

\item{cohort_id_outcome}{cohort_id_outcome}

\item{study_denominator_pop}{study_denominator_pop}

\item{cohort_id_denominator_pop}{cohort_id_denominator_pop}

\item{time_interval}{time_interval}

\item{outcome_washout_window}{outcome_washout_window}

\item{repetitive_events}{repetitive_events}

\item{verbose}{verbose}
}
\description{
Get population incidence estimates
}
