% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pop_prevalence.R
\name{get_pop_prevalence}
\alias{get_pop_prevalence}
\title{Get population prevalence estimates}
\usage{
get_pop_prevalence(
  db,
  results_schema_outcome,
  table_name_outcome,
  cohort_id_outcome = NULL,
  study_denominator_pop,
  cohort_id_denominator_pop = NULL,
  period = "Point",
  time_interval = c("Months"),
  minimum_representative_proportion = 0.5,
  confidence_interval = "exact",
  verbose = FALSE
)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Get population prevalence estimates
}
