% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pop_prevalence.R
\name{get_pop_prevalence}
\alias{get_pop_prevalence}
\title{Get population prevalence estimates}
\usage{
get_pop_prevalence(
  cdm,
  table_name_denominator,
  cohort_id_denominator_pop = NULL,
  table_name_outcome,
  cohort_id_outcome = NULL,
  type = "point",
  time_interval = "months",
  full_period_required = TRUE,
  point = "start",
  minimum_representative_proportion = 0.5,
  verbose = FALSE
)
}
\arguments{
\item{cdm}{CDMConnector CDM reference object}

\item{table_name_denominator}{table_name_denominator}

\item{cohort_id_denominator_pop}{cohort_id_denominator_pop}

\item{table_name_outcome}{table_name_outcome}

\item{cohort_id_outcome}{cohort_id_outcome}

\item{type}{type}

\item{time_interval}{time_interval}

\item{full_period_required}{full period requirement}

\item{point}{point where to compute prevalence inside interval}

\item{minimum_representative_proportion}{minimum_representative_proportion}

\item{verbose}{verbose}
}
\description{
Get population prevalence estimates
}
