% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPopIncidence.R
\name{getPopIncidence}
\alias{getPopIncidence}
\title{Get population incidence estimates}
\usage{
getPopIncidence(
  cdm,
  denominatorTable,
  outcomeTable,
  denominatorId,
  outcomeId,
  interval,
  fullPeriods,
  outcomeWashout,
  repeatedEvents,
  verbose
)
}
\arguments{
\item{cdm}{CDMConnector CDM reference object}

\item{denominatorTable}{denominatorTable}

\item{outcomeTable}{outcomeTable}

\item{denominatorId}{denominatorId}

\item{outcomeId}{outcomeId}

\item{interval}{interval}

\item{fullPeriods}{fullPeriods}

\item{outcomeWashout}{outcomeWashout}

\item{repeatedEvents}{repeatedEvents}

\item{verbose}{verbose}
}
\description{
Get population incidence estimates
}
