% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPopPrevalence.R
\name{getPopPrevalence}
\alias{getPopPrevalence}
\title{Get population prevalence estimates}
\usage{
getPopPrevalence(
  cdm,
  denominatorTable,
  denominatorId,
  outcomeTable,
  outcomeId,
  type,
  interval,
  fullPeriods,
  point,
  fullContribution,
  verbose
)
}
\arguments{
\item{cdm}{CDMConnector CDM reference object}

\item{denominatorTable}{denominatorTable}

\item{denominatorId}{denominatorId}

\item{outcomeTable}{outcomeTable}

\item{outcomeId}{outcomeId}

\item{type}{type}

\item{interval}{interval}

\item{fullPeriods}{full period requirement}

\item{point}{point where to compute prevalence inside interval}

\item{fullContribution}{fullContribution}

\item{verbose}{verbose}
}
\description{
Get population prevalence estimates
}
