% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPopPrevalence.R
\name{getPopPrevalence}
\alias{getPopPrevalence}
\title{Get population prevalence estimates}
\usage{
getPopPrevalence(
  cdm,
  denominatorTable,
  denominatorCohortIds = NULL,
  outcomesTable,
  outcomeCohortIds = NULL,
  type = "point",
  interval = "months",
  fullPeriodsRequired = TRUE,
  point = "start",
  minContribution = 0.5,
  verbose = FALSE
)
}
\arguments{
\item{cdm}{CDMConnector CDM reference object}

\item{denominatorTable}{denominatorTable}

\item{denominatorCohortIds}{denominatorCohortIds}

\item{outcomesTable}{outcomesTable}

\item{outcomeCohortIds}{outcomeCohortIds}

\item{type}{type}

\item{interval}{interval}

\item{fullPeriodsRequired}{full period requirement}

\item{point}{point where to compute prevalence inside interval}

\item{minContribution}{minContribution}

\item{verbose}{verbose}
}
\description{
Get population prevalence estimates
}
