% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obscureCounts.R
\name{obscure_counts}
\alias{obscure_counts}
\title{Obscure the small number of counts}
\usage{
obscure_counts(x, minCellCount = 5, substitute = NA)
}
\arguments{
\item{x}{x}

\item{minCellCount}{minCellCount}

\item{substitute}{substitute}
}
\value{

}
\description{
Obscure the small number of counts
}
