% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{settings}
\alias{settings}
\title{Settings associated with a cohort set}
\usage{
settings(result, ...)
}
\arguments{
\item{result}{Result for which to get settings}

\item{...}{Additional arguments. For denominator cohort, cohortDefinitionId
can be used to specify a particular cohort of interest. For incidence and
prevalence analyses results, analysisId can be used to specify a
particular analysis of interest.}
}
\value{
tibble with settings used when generating the cohort set
}
\description{
Settings associated with a cohort set
}
\examples{
cdm <- mockIncidencePrevalenceRef(sampleSize = 200)
cdm$denominator <- generateDenominatorCohortSet(cdm,
  startDate = NULL,
  endDate = NULL,
  ageGroup = list(c(18, 40)),
  sex = c("Female", "Male"),
  daysPriorHistory = 120
)
settings(result = cdm$denominator)
settings(result = cdm$denominator, cohortDefinitionId = 1)
incidence <- estimateIncidence(
  cdm = cdm,
  denominatorTable = "denominator",
  outcomeTable = "outcome",
  interval = "overall"
)
settings(result = incidence)
settings(result = incidence, analysisId = 1)
}
