% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{settings}
\alias{settings}
\title{Settings associated with a cohort set}
\usage{
settings(x)
}
\arguments{
\item{x}{Cohort set for which to get settings}
}
\value{
tibble with settings used when generating the cohort set
}
\description{
Settings associated with a cohort set
}
