% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tableIncidence}
\alias{tableIncidence}
\title{Table of incidence results}
\usage{
tableIncidence(result, type = "gt", .options = list())
}
\arguments{
\item{result}{Incidence results}

\item{type}{Type of table. Can be "gt", "flextable", or "tibble".}

\item{.options}{Table options to apply}
}
\value{
Table of results
}
\description{
Table of incidence results
}
