% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tablePrevalence}
\alias{tablePrevalence}
\title{Format a point_prevalence object into a visual table.}
\usage{
tablePrevalence(
  result,
  prevalenceType,
  formatEstimateName = c(`Denominator (N)` = "<denominator_count>", `Outcome (N)` =
    "<outcome_count>", `Prevalence [95\% CI]` =
    "<prevalence> (<prevalence_95CI_lower> - <prevalence_95CI_upper>)"),
  header = c("group", "strata"),
  splitStrata = TRUE,
  cdmName = TRUE,
  outcomeName = TRUE,
  outcomeSettings = FALSE,
  denominatorName = TRUE,
  denominatorSettings = FALSE,
  analysisSettings = FALSE,
  groupColumn = NULL,
  type = "gt",
  .options = list()
)
}
\arguments{
\item{result}{A summarised_result object with results from
estimatePointPrevalence().}

\item{prevalenceType}{Type of prevalence estimates: "point" or "period".}

\item{formatEstimateName}{Named list of estimate name's to join, sorted by
computation order. Indicate estimate_name's between <...>.}

\item{header}{A vector containing which elements should go into the header
in order. Allowed are: \code{cdm_name}, \code{group}, \code{strata}, \code{additional},
\code{variable}, \code{estimate}, \code{settings}.}

\item{splitStrata}{If TRUE strata columns will be splitted.}

\item{cdmName}{If TRUE database names will be displayed.}

\item{outcomeName}{If TRUE outcome cohort names will be displayed.}

\item{outcomeSettings}{If TRUE settings related to the outcome cohorts will
be displayed.}

\item{denominatorName}{If TRUE denominator cohort names will be displayed.}

\item{denominatorSettings}{If TRUE settings related to the denominator cohorts
will be displayed.}

\item{analysisSettings}{If TRUE database names will be displayed.}

\item{groupColumn}{Column to use as group labels.
#' @param type Type of desired formatted table, possibilities: "gt",
"flextable", "tibble".}

\item{.options}{Named list with additional formatting options.
IncidencePrevalence::optionsTablePrevalence() shows allowed
arguments and their default values.}
}
\value{
A table with a formatted version of the point prevalence result.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\donttest{
library(IncidencePrevalence)

cdm <- mockIncidencePrevalenceRef()

cdm <- generateDenominatorCohortSet(cdm = cdm, name = "denominator")

prev <- estimatePointPrevalence(
  cdm = cdm,
  denominatorTable = "denominator",
  outcomeTable = "outcome",
  summarisedResult = TRUE,
)

tablePrevalence(prev, prevalenceType = "point")

CDMConnector::cdmDisconnect(cdm = cdm)
}

}
