% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tablePrevalenceAttrition}
\alias{tablePrevalenceAttrition}
\title{Table of prevalence attrition results}
\usage{
tablePrevalenceAttrition(
  result,
  type = "gt",
  header = "variable_name",
  groupColumn = c("cdm_name", "variable_level"),
  settingsColumns = colnames(settings(result)),
  hide = "estimate_name"
)
}
\arguments{
\item{result}{A summarised_result object. Output of
summariseCohortAttrition().}

\item{type}{Type of table. Check supported types with
\code{visOmopResults::tableType()}.}

\item{header}{Columns to use as header. See options with
\code{colnames(visOmopResults::splitAll(result))}. Variables in \code{settingsColumns}
are also allowed}

\item{groupColumn}{Variables to use as group labels. Allowed columns are the
same as in \code{header}}

\item{settingsColumns}{Variables from the settings atribute to dispaly in
the table}

\item{hide}{Table columns to exclude, options are the ones described in
\code{header}}
}
\value{
A visual table.
}
\description{
Table of prevalence attrition results
}
